Fix #81: blocking delay for unreachable remote
This patch fixes a blocking delay when a named remote action is set up.
On GLIBC systems, with the default resolver setup, this may be as long
as 10 seconds, meaning syslogd retried nslookup() every log message,
which in turn lead to snail speed on logging to other targets.

Additionally it re-enables the warning for failure to resolve the DNS
name -- a user is likely very interested in why they are not seeing any
logs on their remote server.  Logs age out, so logging in to a device
which may have logged this once is not very useful.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>