Fix #82: retry creating UNIX and network sockets on failure
When starting up, or on SIGHUP, syslogd may fail to create a UNIX socket
or (bind) an Internet socket.  The latter may be because the system is
in bootstrap and the interface and/or address syslogd tries to bind() to
does not yet exist.

This commit adds a very basic retry loop every five seconds.

Also, with the new retry logic it was deemed worth it to increase the
logging from syslogd itself.  I.e., when failing to create a socket,
or when retrying and succeeding, as well as when closing sockets.  In
particular when opening and closing Internet facing ports.

Since startup, or reconfiguration, is not failure to send() the we do
not want to inadvertedly trigger F_FORW_SUSP for remote log actions.
This is the reason for the ENONET error code from socket_ffs(), to
ensure we retry sending to remote on the next log message instead of
waiting 180 sec.  Ordering at boot, or reconf, is also the reason for
setting Initialized earlier -- this prevents inadvertedly logging to the
console instead of to a file that actually is open.  We can usually log
to file but maybe not yet to a remote.

To not terrify casual readers of the logs, messages for opening and
closing Inet (Internet) sockets are tracked from what context they were
opened in.  Context, in this context, is the secure_mode switch in which
syslogd only opens UDP sockets for sending messages to other syslog
servers, i.e., without binding to a server port.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>